Created Input Manager.
Took a while to learn how to use the new Input System. Seems pretty flexible and useful. If anyone is interested, I created a script sample on how to use the Input Manager.

Ping me on Discord if you guys want me to make a script sample on how to use certain systems.