Discovered "using static" and "sealed " keywords
"using static..." is useful in regards to simplify the calling of enumerators or even inner classes. It goes from GameManager.InputManager to InputManager. Though it won't work if the static variable name inside the class is the same as the class name. Maybe we should consider refactoring them to camelCase.

'public sealed class...' is an interesting keyword where upon using it, it prevents classes from inheriting from it. It is recommended to just spam it on every class you see and just remove it if the class needs to be inherited from. In doing so, it can apparently dramatically increase the speed of code.