Scan ability implemented (Read desc)
The scan ability fully works now. There is one thing that needs to be done with the objective icons and that is to update the objective prefabs by removing the objective icon and adding in a ObjectiveIcon prefab onto it. The prefab has a Transform serialized field, just grab the transform of the parent the prefab is attached to. Look at the terminal prefab for reference.